requirements.txt
pandas>=1.5.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
tqdm>=4.64.0
python-dotenv>=0.19.0
apache-airflow>=2.7.0
numpy>=1.21.0
